requirements.txt
Django==5.0
djangorestframework==3.15
psycopg[binary]==3.1.20
redis==5.0
django-cors-headers==4.0
drf-spectacular==0.28.0
pillow==11.3
drf_yasg